Move validator models into context namespace
