+++ src/InvestProvider.Backend/Services/Handlers/ContextBuilders/PhaseContextBuilder.cs
@@ -1,7 +1,7 @@
 using Amazon.DynamoDBv2.DataModel;
 using InvestProvider.Backend.Services.DynamoDb.Models;
 using InvestProvider.Backend.Services.Strapi;
-using InvestProvider.Backend.Services.Validators.Models;
+using InvestProvider.Backend.Services.Handlers.Contexts;
 
 namespace InvestProvider.Backend.Services.Handlers.ContextBuilders;
