+++ src/InvestProvider.Backend/Services/Handlers/MyAllocation/Models/MyAllocationRequest.cs
@@ -4,7 +4,7 @@ using Net.Web3.EthereumWallet;
 using Poolz.Finance.CSharp.Strapi;
 using InvestProvider.Backend.Services.Strapi.Models;
 using InvestProvider.Backend.Services.DynamoDb.Models;
-using InvestProvider.Backend.Services.Validators.Models;
+using InvestProvider.Backend.Services.Handlers.Contexts;
 using ProjectsInformation = InvestProvider.Backend.Services.DynamoDb.Models.ProjectsInformation;
 
 namespace InvestProvider.Backend.Services.Handlers.MyAllocation.Models;
