--- src/InvestProvider.Backend/Services/Validators/Models/IExistActivePhase.cs
@@ -1,9 +1,0 @@
-﻿using InvestProvider.Backend.Services.Strapi.Models;
-
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IExistActivePhase : IProjectContext
-{
-    bool FilterPhases { get; }
-    ProjectInfo StrapiProjectInfo { get; set; }
-}
