--- src/InvestProvider.Backend/Services/Validators/Models/IExistPhase.cs
@@ -1,8 +1,0 @@
-﻿using Poolz.Finance.CSharp.Strapi;
-
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IExistPhase : IExistActivePhase
-{
-    ComponentPhaseStartEndAmount Phase { get; set; }
-}
