--- src/InvestProvider.Backend/Services/Validators/Models/IInvestPool.cs
@@ -1,6 +1,0 @@
-﻿namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IInvestPool : IProjectContext
-{
-    public long PoolzBackId { get; }
-}
