--- src/InvestProvider.Backend/Services/Validators/Models/IProjectContext.cs
@@ -1,8 +1,0 @@
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IProjectContext
-{
-    string ProjectId { get; }
-    string PhaseId { get; }
-    long ChainId { get; }
-}
