--- src/InvestProvider.Backend/Services/Validators/Models/IValidatedDynamoDbProjectInfo.cs
@@ -1,8 +1,0 @@
-﻿using InvestProvider.Backend.Services.DynamoDb.Models;
-
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IValidatedDynamoDbProjectInfo : IProjectContext
-{
-    ProjectsInformation DynamoDbProjectsInfo { get; set; }
-}
