--- src/InvestProvider.Backend/Services/Validators/Models/IWhiteListPhase.cs
@@ -1,8 +1,0 @@
-﻿using Poolz.Finance.CSharp.Strapi;
-
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IWhiteListPhase
-{
-    public ComponentPhaseStartEndAmount Phase { get; }
-}
