--- src/InvestProvider.Backend/Services/Validators/Models/IWhiteListUser.cs
@@ -1,10 +1,0 @@
-﻿using InvestProvider.Backend.Services.DynamoDb.Models;
-using Net.Web3.EthereumWallet;
-
-namespace InvestProvider.Backend.Services.Validators.Models;
-
-public interface IWhiteListUser : IExistActivePhase
-{
-    public WhiteList WhiteList { get; set; }
-    public EthereumAddress UserAddress { get; }
-}
